requirements.txt
PyOpenGL 
PyOpenGL_accelerate
numpy
scipy
matplotlib
pyqt5
#python3.6